One more cleanup pass
