--- frontend/src/components/pages/rp-connect/utils/component-aliases.test.ts
@@ -1,48 +1,0 @@
-/**
- * Copyright 2026 Redpanda Data, Inc.
- *
- * Use of this software is governed by the Business Source License
- * included in the file https://github.com/redpanda-data/redpanda/blob/dev/licenses/bsl.md
- *
- * As of the Change Date specified in that file, in accordance with
- * the Business Source License, use of this software will be governed
- * by the Apache License, Version 2.0
- */
-
-import { describe, expect, it } from 'vitest';
-
-import { aliasTermsForName, COMPONENT_ALIASES } from './component-aliases';
-
-describe('component aliases', () => {
-  it('maps intent terms to a component name via its fragments', () => {
-    // "queue" lists kafka, so kafka_franz surfaces under that intent.
-    expect(aliasTermsForName('kafka_franz')).toContain('queue');
-    expect(aliasTermsForName('kafka_franz')).toContain('stream');
-  });
-
-  it('maps transform/map intents to bloblang/mapping', () => {
-    expect(aliasTermsForName('mapping')).toEqual(expect.arrayContaining(['transform', 'map']));
-    expect(aliasTermsForName('bloblang')).toContain('transform');
-  });
-
-  it('returns no alias terms for a name no fragment matches', () => {
-    expect(aliasTermsForName('totally_unrelated_xyz')).toEqual([]);
-  });
-
-  it('matches fragments on `_`/word boundaries, not mid-token substrings', () => {
-    // Whole multi-token fragments and `_`-delimited tokens match...
-    expect(aliasTermsForName('rate_limit')).toEqual(expect.arrayContaining(['throttle', 'ratelimit']));
-    expect(aliasTermsForName('json_schema')).toContain('json');
-    // ...but a fragment buried mid-token doesn't (`log` inside `catalog`).
-    expect(aliasTermsForName('catalog')).not.toContain('log');
-  });
-
-  it('keeps alias keys and fragments lowercase', () => {
-    for (const [key, fragments] of Object.entries(COMPONENT_ALIASES)) {
-      expect(key).toBe(key.toLowerCase());
-      for (const fragment of fragments) {
-        expect(fragment).toBe(fragment.toLowerCase());
-      }
-    }
-  });
-});
